Add files via upload
Arduino Uno is to be flashed with sketch_oct10a.ino
Assemble the circuit as shown in the uploaded PDF. In the project, LM358D and IN4007 diodes were used. A simple sine wave was obtained by passing the output from capacitor in 555 IC astable multi-vibrator through another capacitor connected to ground (it acted as an integrator). 
PESMOaq.py was used to read the data from serial port and plot it.